refactor(transport): rework the abort helpers
---
